+++ crates/bitcoin-da/src/service.rs
@@ -20,7 +20,6 @@ use bitcoin::{Amount, BlockHash, CompactTarget, Transaction, Txid, Wtxid};
 use bitcoincore_rpc::json::TestMempoolAcceptResult;
 use bitcoincore_rpc::{Auth, Client, Error, RpcApi, RpcError};
 use borsh::BorshDeserialize;
-use citrea_common::FromEnv;
 use serde::{Deserialize, Serialize};
 use sov_rollup_interface::da::{DaData, DaDataBatchProof, DaDataLightClient, DaSpec};
 use sov_rollup_interface::services::da::{DaService, SenderWithNotifier};
@@ -85,18 +84,7 @@ pub struct BitcoinServiceConfig {
     // absolute path to the directory where the txs will be written to
     pub tx_backup_dir: String,
 }
-impl FromEnv for BitcoinServiceConfig {
-    fn from_env() -> anyhow::Result<Self> {
-        Ok(Self {
-            node_url: std::env::var("NODE_URL")?,
-            node_username: std::env::var("NODE_USERNAME")?,
-            node_password: std::env::var("NODE_PASSWORD")?,
-            network: serde_json::from_str(&format!("\"{}\"", std::env::var("NETWORK")?))?,
-            da_private_key: std::env::var("DA_PRIVATE_KEY").ok(),
-            tx_backup_dir: std::env::var("TX_BACKUP_DIR")?,
-        })
-    }
-}
+
 /// A service that provides data and data availability proofs for Bitcoin
 #[derive(Debug)]
 pub struct BitcoinService {
