+++ crates/common/Cargo.toml
@@ -30,10 +30,12 @@ tracing = { workspace = true }
 
 # Sov SDK deps
 sov-db = { path = "../sovereign-sdk/full-node/db/sov-db" }
+sov-mock-da = { path = "../sovereign-sdk/adapters/mock-da" }
 sov-rollup-interface = { path = "../sovereign-sdk/rollup-interface" }
 sov-stf-runner = { path = "../sovereign-sdk/full-node/sov-stf-runner", features = ["native"] }
 
 # Citrea
+bitcoin-da = { path = "../bitcoin-da", features = ["native"] }
 citrea-primitives = { path = "../primitives/" }
 citrea-pruning = { path = "../pruning" }
